+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 
 ## New Features
 
-- PR #41 jaccard grmat graph bindings
+- PR #73 Weighted Jaccard bindings
+- PR #41 RMAT graph bindings
 - PR #43 Louvain binings
 - PR #44 SSSP bindings
+- PR #47 BSF bindings
+- PR #53 New Repo structure
+- PR #67 RMM Integration with rmm as as submodule
 
 ## Improvements
 
@@ -14,6 +18,7 @@
 ## Bug Fixes
 
 - PR #48 ABI Fixes
+- PR #72 Bug fix for segfault issue getting transpose from adjacency list
 
 
 # cuGraph 0.5.0 (28 Jan 2019)